pyproject.toml
[project]
name = "projeto-melhoramais"
version = "0.1.0"
description = ""
authors = [
    {name = "Kauan Massuia",email = "kauan.massuia@sou.inteli.edu.br"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas (>=3.0.1,<4.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
